deployment/DockerFiles/requirements.txt
torch
transformers
gradio
joblib
pillow
